Move SensorAPI to API package under dev version
