+++ flexmeasures/api/__init__.py
@@ -105,9 +105,11 @@ def register_at(app: Flask):
     from flexmeasures.api.v1_2 import register_at as v1_2_register_at
     from flexmeasures.api.v1_3 import register_at as v1_3_register_at
     from flexmeasures.api.v2_0 import register_at as v2_0_register_at
+    from flexmeasures.api.dev import register_at as dev_register_at
 
     v1_register_at(app)
     v1_1_register_at(app)
     v1_2_register_at(app)
     v1_3_register_at(app)
     v2_0_register_at(app)
+    dev_register_at(app)
